Counter: Fix compiler warning about signed-unsigned comparison

+++ src/_counter.h
@@ -24,6 +24,7 @@
 #ifndef PCT__COUNTER_H
 #define PCT__COUNTER_H
 
+#include "Python.h"
 #include "config.h"
 #if HAVE_STDINT_H
 # include <stdint.h>
@@ -38,7 +39,7 @@ typedef struct {
     PyBytesObject *prefix;     /* Prefix (useful for a nonce) */
     PyBytesObject *suffix;     /* Suffix (useful for a nonce) */
     uint8_t *val;       /* Buffer for our output string */
-    uint32_t buf_size;  /* Size of the buffer */
+    Py_ssize_t buf_size;/* Size of the buffer */
     uint8_t *p;         /* Pointer to the part of the buffer that we're allowed to update */
     uint16_t nbytes;    /* The number of bytes that from .p that are part of the counter */
     void (*inc_func)(void *);   /* Pointer to the counter increment function */
